requirements.txt
google-api-python-client==2.8.0
google-auth-httplib2==0.1.0
boto3==1.17.79
requests==2.25.1